server/complete_routing_app/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pillow==10.0.0
customtkinter==5.2.0
geopy==2.4.0
folium==0.14.0 